requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
Brotli==1.1.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
fastapi==0.110.0
frozenlist==1.4.1
g4f==0.2.7.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.5
openai==1.14.3
OpenAIAuth==3.0.0
prompt-toolkit==3.0.43
pycryptodome==3.20.0
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.17.2
PySocks==1.7.1
python-dotenv==1.0.1
regex==2023.12.25
requests==2.31.0
revChatGPT==6.8.6
rich==13.7.1
sniffio==1.3.1
socksio==1.0.0
SQLAlchemy==2.0.29
starlette==0.36.3
tiktoken==0.6.0
tls-client==1.0.1
tqdm==4.66.2
typing_extensions==4.10.0
urllib3==2.2.1
uvicorn==0.29.0
wcwidth==0.2.13
yarl==1.9.4
beautifulsoup4~=4.12.3
pillow~=10.3.0